refactor(library/URLComponent): moves "Path.ts" to "Path/definition.ts"
